requirements.txt
django == 4.1
django-rest-framework
django-crispy-forms